requirements.txt
pyyaml>=6.0.1
playwright>=1.49.0
requests>=2.31.0
